# chenoj-frontend/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "接口文档",
    "contact": {}
  },
  "host": "localhost:8121",
  "basePath": "/api",
...
  "paths": {
...
    "/api/post/list/page": {
      "post": {
        "tags": [
          "post-controller"
        ],
        "summary": "listPostByPage",
        "operationId": "listPostByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "PostQueryRequest",
              "$ref": "#/definitions/PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«Post»»",
              "$ref": "#/definitions/BaseResponse«Page«Post»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«Post»»",
              "$ref": "#/definitions/BaseResponse«Page«Post»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/api/post_favour/": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "doPostFavour",
        "operationId": "doPostFavourUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postFavourAddRequest",
            "description": "postFavourAddRequest",
            "required": true,
            "schema": {
              "originalRef": "PostFavourAddRequest",
              "$ref": "#/definitions/PostFavourAddRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«int»",
              "$ref": "#/definitions/BaseResponse«int»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«int»",
              "$ref": "#/definitions/BaseResponse«int»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/post_favour/list/page": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "listFavourPostByPage",
        "operationId": "listFavourPostByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postFavourQueryRequest",
            "description": "postFavourQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "PostFavourQueryRequest",
              "$ref": "#/definitions/PostFavourQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«PostVO»»",
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«PostVO»»",
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/post_favour/my/list/page": {
      "post": {
        "tags": [
          "post-favour-controller"
        ],
        "summary": "listMyFavourPostByPage",
        "operationId": "listMyFavourPostByPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "postQueryRequest",
            "description": "postQueryRequest",
            "required": true,
            "schema": {
              "originalRef": "PostQueryRequest",
              "$ref": "#/definitions/PostQueryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«PostVO»»",
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«Page«PostVO»»",
              "$ref": "#/definitions/BaseResponse«Page«PostVO»»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/api/user/get/login": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "getLoginUser",
        "operationId": "getLoginUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "BaseResponse«LoginUserVO»",
              "$ref": "#/definitions/BaseResponse«LoginUserVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
  },
  "definitions": {
...
    "BaseResponse«LoginUserVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "LoginUserVO",
          "$ref": "#/definitions/LoginUserVO"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«LoginUserVO»"
    },
    "BaseResponse«Page«PostVO»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "Page«PostVO»",
          "$ref": "#/definitions/Page«PostVO»"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Page«PostVO»»"
    },
    "BaseResponse«Page«Post»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "Page«Post»",
          "$ref": "#/definitions/Page«Post»"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«Page«Post»»"
    },
...
    "BaseResponse«int»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "BaseResponse«int»"
    },
...
    "LoginUserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "LoginUserVO"
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean"
        },
        "column": {
          "type": "string"
        }
      },
      "title": "OrderItem"
    },
    "Page«PostVO»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "originalRef": "OrderItem",
            "$ref": "#/definitions/OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "originalRef": "PostVO",
            "$ref": "#/definitions/PostVO"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«PostVO»"
    },
    "Page«Post»": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int64"
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64"
        },
        "optimizeCountSql": {
          "type": "boolean"
        },
        "orders": {
          "type": "array",
          "items": {
            "originalRef": "OrderItem",
            "$ref": "#/definitions/OrderItem"
          }
        },
        "pages": {
          "type": "integer",
          "format": "int64"
        },
        "records": {
          "type": "array",
          "items": {
            "originalRef": "Post",
            "$ref": "#/definitions/Post"
          }
        },
        "searchCount": {
          "type": "boolean"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Page«Post»"
    },
...
    "Post": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "isDelete": {
          "type": "integer",
          "format": "int32"
        },
        "tags": {
          "type": "string"
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "Post"
    },
...
    "PostFavourAddRequest": {
      "type": "object",
      "properties": {
        "postId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostFavourAddRequest"
    },
    "PostFavourQueryRequest": {
      "type": "object",
      "properties": {
        "current": {
          "type": "integer",
          "format": "int32"
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "postQueryRequest": {
          "originalRef": "PostQueryRequest",
          "$ref": "#/definitions/PostQueryRequest"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostFavourQueryRequest"
    },
    "PostQueryRequest": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "current": {
          "type": "integer",
          "format": "int32"
        },
        "favourUserId": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "notId": {
          "type": "integer",
          "format": "int64"
        },
        "orTags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pageSize": {
          "type": "integer",
          "format": "int32"
        },
        "searchText": {
          "type": "string"
        },
        "sortField": {
          "type": "string"
        },
        "sortOrder": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostQueryRequest"
    },
...
    "PostVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "favourNum": {
          "type": "integer",
          "format": "int32"
        },
        "hasFavour": {
          "type": "boolean"
        },
        "hasThumb": {
          "type": "boolean"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "tagList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "thumbNum": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "user": {
          "originalRef": "UserVO",
          "$ref": "#/definitions/UserVO"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "PostVO"
    },
...
    "UserVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userAvatar": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "userProfile": {
          "type": "string"
        },
        "userRole": {
          "type": "string"
        }
      },
      "title": "UserVO"
    }
  },
...
}